Add more compilation flags (#3638)

+++ src/engine/serialize.cpp
@@ -222,7 +222,7 @@ StreamBase & StreamBase::operator<<( const float v )
 
 StreamBase & StreamBase::operator<<( const std::string & v )
 {
-    put32( v.size() );
+    put32( static_cast<uint32_t>( v.size() ) );
 
     for ( std::string::const_iterator it = v.begin(); it != v.end(); ++it )
         put8( *it );
